rest-validation-3.1/src/main/java/net/petrikainulainen/spring/trenches/comment/dto/ValidationErrorDTO.java
package net.petrikainulainen.spring.trenches.comment.dto;

import java.util.ArrayList;
import java.util.List;

/**
 * @author Petri Kainulainen
 */
public class ValidationErrorDTO {

    private List<FieldErrorDTO> fieldErrors = new ArrayList<>();

    public ValidationErrorDTO() {

    }

    public void addFieldError(String path, String message) {
        FieldErrorDTO error = new FieldErrorDTO(path, message);
        fieldErrors.add(error);
    }

    public List<FieldErrorDTO> getFieldErrors() {
        return fieldErrors;
    }
}
